pyproject.toml
[project]
name = "nl-ai-retail-lakehouse-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.32",
    "ollama (>=0.6.1,<0.7.0)",
    "streamlit (>=1.56.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "retail_lakehouse"}]

[dependency-groups]
dev = [
    "isort (>=8.0.1,<9.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "sqlfluff (>=4.1.0,<5.0.0)"
]
